Address Copilot review on the thumbnail grid
Read thumbnailMaxExtent once per layout instead of once per tile: it
scans every entry, so reading it inside the ForEach walked the whole
grid for each tile in it.

Cast both of PieceSpan's axes explicitly in extent(span:puzzleAspect:).
It compiled as it was — CGSize has a Double initializer — but casting
one axis and not the other invited the question.

Correct the tile's rotation comment. The drawn bounds rest within the
square at either end of a quarter turn, but sweep wider mid-turn, so a
near-square piece at the grid's largest scale can graze the clip while
it animates. Sizing every piece down to clear its own diagonal would
cost the whole grid real size to buy a transient on one tile, so the
animation stays.

+++ ios/Pussel/Features/Solve/PieceThumbnailGeometry.swift
@@ -21,7 +21,7 @@ enum PieceThumbnailGeometry {
   /// (pre-rotation) axes, matching what `PieceSpan` describes.
   static func extent(span: PieceSpan, puzzleAspect: CGFloat) -> CGSize {
     guard puzzleAspect > 0 else { return .zero }
-    return CGSize(width: span.width, height: CGFloat(span.height) / puzzleAspect)
+    return CGSize(width: CGFloat(span.width), height: CGFloat(span.height) / puzzleAspect)
   }
 
   /// The largest single dimension any measured piece reaches, in puzzle-width
